+++ HISTORY.rst
@@ -15,6 +15,7 @@ Recent History
 Next Release
 ------------------
 
+* Add --transfer command line option to transfer selected Meetup events.
 * Allow multiple Meetup event IDs with a --skip command line option.
 
 1.6.0 (2021-03-24)